Add new service section and image to servicios page
Introduced a new section highlighting reasons to work with the company and added a call-to-action button on the servicios page. Included a new image (funco-de-pie-en-oficina.png) and updated styles in app.css to support new layout and responsive design. Updated general.tsx to export the new image and made minor adjustments to the contacto page title.